requirements.txt
rich 
tqdm 
requests 
onvif_zeep 
#miniupnpc 
wsdiscovery 
jinja2
